refactor(plugins): split two self-contained pieces out of PluginStoreSetup.kt
PluginStoreSetup.kt is a 2,089-line god object owning six unrelated
concerns (#447). Rather than attempt the full eight-step refactor in one
pass, this extracts the two pieces that are safely, mechanically
verifiable without a live app run:

- SidecarBackfillCoordinator (step 7): already a standalone class driven
  entirely by constructor lambdas, with zero references to
  PluginStoreSetup's own state. Moved to its own file in the same
  package, so no caller or test needed to change - a pure file move,
  not a behavior change.
- isNewerVersion, isTooOldForHost, pickPluginJarUrl and
  extractVersionFromJarFileName (step 6): four pure, stateless functions
  already `internal` for test access. Extracted into a new
  PluginVersionComparator object; the ~9 internal call sites and the
  test file (renamed PluginVersionComparatorTest, same assertions) were
  updated to match.

PluginStoreSetup.kt drops from 2,089 to 1,894 lines. The remaining six
steps (GitHubReleaseClient, SystemPluginInstaller, PluginSignatureResolver,
BundledPluginCopier, readPluginManifest relocation, and the full
initialize()/loadPersistedPlugins()/shutdown() orchestrator shape) are
not attempted here - several of them touch the plugin-install and
startup-restore paths the issue's own acceptance criteria says to verify
via `./gradlew run`, which this environment cannot do.
