+++ composeApp/src/desktopTest/kotlin/ai/rever/boss/plugin/PluginVersionComparatorTest.kt
@@ -8,75 +8,79 @@ import kotlin.test.assertTrue
 
 /**
  * Guards the system-plugin minVersion gate (editor-tab 1.4.0 bundling
- * BossEditor after the host dropped it): [PluginStoreSetup.isTooOldForHost]
+ * BossEditor after the host dropped it): [PluginVersionComparator.isTooOldForHost]
  * decides whether the installed JAR must be replaced before load,
- * [PluginStoreSetup.isNewerVersion] is the comparator behind it (including
- * the release-vs-pre-release tiebreak), and [PluginStoreSetup.pickPluginJarUrl]
+ * [PluginVersionComparator.isNewerVersion] is the comparator behind it (including
+ * the release-vs-pre-release tiebreak), and [PluginVersionComparator.pickPluginJarUrl]
  * must never select a "-thin.jar" release asset (a module's default :jar
  * output, missing everything buildPluginJar bundles).
+ *
+ * Relocated from PluginStoreSetupMinVersionGateTest (BossConsole#447 step 6) when
+ * the four functions it exercises moved into their own file - same assertions,
+ * new home.
  */
-class PluginStoreSetupMinVersionGateTest {
+class PluginVersionComparatorTest {
     // ---- isTooOldForHost: the load/update decision ----
 
     @Test
     fun `no minimum means never too old`() {
-        assertFalse(PluginStoreSetup.isTooOldForHost("0.0.1", null))
-        assertFalse(PluginStoreSetup.isTooOldForHost(null, null))
+        assertFalse(PluginVersionComparator.isTooOldForHost("0.0.1", null))
+        assertFalse(PluginVersionComparator.isTooOldForHost(null, null))
     }
 
     @Test
     fun `version equal to minimum loads`() {
-        assertFalse(PluginStoreSetup.isTooOldForHost("1.4.0", "1.4.0"))
+        assertFalse(PluginVersionComparator.isTooOldForHost("1.4.0", "1.4.0"))
     }
 
     @Test
     fun `version above minimum loads`() {
-        assertFalse(PluginStoreSetup.isTooOldForHost("1.4.1", "1.4.0"))
-        assertFalse(PluginStoreSetup.isTooOldForHost("2.0.0", "1.4.0"))
+        assertFalse(PluginVersionComparator.isTooOldForHost("1.4.1", "1.4.0"))
+        assertFalse(PluginVersionComparator.isTooOldForHost("2.0.0", "1.4.0"))
     }
 
     @Test
     fun `version below minimum forces update`() {
-        assertTrue(PluginStoreSetup.isTooOldForHost("1.3.2", "1.4.0"))
+        assertTrue(PluginVersionComparator.isTooOldForHost("1.3.2", "1.4.0"))
     }
 
     @Test
     fun `unreadable installed version forces update`() {
         // Only very old JARs lack a readable version — conservatively too old.
-        assertTrue(PluginStoreSetup.isTooOldForHost(null, "1.4.0"))
+        assertTrue(PluginVersionComparator.isTooOldForHost(null, "1.4.0"))
     }
 
     @Test
     fun `pre-release of the minimum forces update`() {
         // 1.4.0-rc1 predates 1.4.0 and may lack the final contract.
-        assertTrue(PluginStoreSetup.isTooOldForHost("1.4.0-rc1", "1.4.0"))
+        assertTrue(PluginVersionComparator.isTooOldForHost("1.4.0-rc1", "1.4.0"))
     }
 
     // ---- isNewerVersion: comparator edges ----
 
     @Test
     fun `basic semver ordering`() {
-        assertTrue(PluginStoreSetup.isNewerVersion("1.4.0", "1.3.9"))
-        assertFalse(PluginStoreSetup.isNewerVersion("1.3.9", "1.4.0"))
-        assertFalse(PluginStoreSetup.isNewerVersion("1.4.0", "1.4.0"))
+        assertTrue(PluginVersionComparator.isNewerVersion("1.4.0", "1.3.9"))
+        assertFalse(PluginVersionComparator.isNewerVersion("1.3.9", "1.4.0"))
+        assertFalse(PluginVersionComparator.isNewerVersion("1.4.0", "1.4.0"))
     }
 
     @Test
     fun `multi-digit segments compare numerically not lexically`() {
-        assertTrue(PluginStoreSetup.isNewerVersion("1.4.10", "1.4.2"))
-        assertFalse(PluginStoreSetup.isNewerVersion("1.4.2", "1.4.10"))
+        assertTrue(PluginVersionComparator.isNewerVersion("1.4.10", "1.4.2"))
+        assertFalse(PluginVersionComparator.isNewerVersion("1.4.2", "1.4.10"))
     }
 
     @Test
     fun `release is newer than its own pre-release`() {
-        assertTrue(PluginStoreSetup.isNewerVersion("1.4.0", "1.4.0-rc1"))
-        assertFalse(PluginStoreSetup.isNewerVersion("1.4.0-rc1", "1.4.0"))
+        assertTrue(PluginVersionComparator.isNewerVersion("1.4.0", "1.4.0-rc1"))
+        assertFalse(PluginVersionComparator.isNewerVersion("1.4.0-rc1", "1.4.0"))
     }
 
     @Test
     fun `suffixed segment counts as its numeric prefix`() {
         // 1.4.10-beta is numerically 1.4.10, which beats 1.4.2.
-        assertTrue(PluginStoreSetup.isNewerVersion("1.4.10-beta", "1.4.2"))
+        assertTrue(PluginVersionComparator.isNewerVersion("1.4.10-beta", "1.4.2"))
     }
 
     // ---- pickPluginJarUrl: release-asset selection ----
@@ -95,20 +99,20 @@ class PluginStoreSetupMinVersionGateTest {
             )
         assertEquals(
             "https://github.example/boss-plugin-editor-tab-1.4.3.jar",
-            PluginStoreSetup.pickPluginJarUrl(json, "boss-plugin-editor-tab"),
+            PluginVersionComparator.pickPluginJarUrl(json, "boss-plugin-editor-tab"),
         )
     }
 
     @Test
     fun `release with only a thin jar yields null`() {
         val json = releaseJson("https://github.example/boss-plugin-editor-tab-1.4.3-thin.jar")
-        assertNull(PluginStoreSetup.pickPluginJarUrl(json, "boss-plugin-editor-tab"))
+        assertNull(PluginVersionComparator.pickPluginJarUrl(json, "boss-plugin-editor-tab"))
     }
 
     @Test
     fun `release without matching assets yields null`() {
         val json = releaseJson("https://github.example/some-other-plugin-1.0.0.jar")
-        assertNull(PluginStoreSetup.pickPluginJarUrl(json, "boss-plugin-editor-tab"))
+        assertNull(PluginVersionComparator.pickPluginJarUrl(json, "boss-plugin-editor-tab"))
     }
 
     // ---- extractVersionFromJarFileName: version source for the gate ----
@@ -117,7 +121,7 @@ class PluginStoreSetupMinVersionGateTest {
     fun `version extracted from standard jar filename`() {
         assertEquals(
             "1.4.0",
-            PluginStoreSetup.extractVersionFromJarFileName(
+            PluginVersionComparator.extractVersionFromJarFileName(
                 "boss-plugin-editor-tab-1.4.0.jar",
                 "boss-plugin-editor-tab",
             ),
@@ -127,7 +131,7 @@ class PluginStoreSetupMinVersionGateTest {
     @Test
     fun `non-matching filename yields null`() {
         assertNull(
-            PluginStoreSetup.extractVersionFromJarFileName(
+            PluginVersionComparator.extractVersionFromJarFileName(
                 "other-plugin-1.4.0.jar",
                 "boss-plugin-editor-tab",
             ),
